redefine markdown to other class
