+++ core/chunking/chunkers/chunker.py
@@ -45,27 +45,3 @@ def token(size,input_text):
     logger.info("provided chunk size : "+str(size))
     logger.info("First chunk : "+chunks_text[0])
     return chunks_text    
-
-
-def markdown_header(input_text):
-    
-    headers_to_split_on = [
-    ("h1", "Header 1"),
-    ("h2", "Header 2"),
-    ("h3", "Header 3"),
-    ]
-
-    html_splitter = HTMLHeaderTextSplitter(headers_to_split_on=headers_to_split_on)
-    html_header_splits = html_splitter.split_text(input_text)
-    return headers_to_split_on, html_header_splits
-    
-    
-def markdown_section(input_text):
-    
-    headers_to_split_on = [("h1", "Header 1"), ("h2", "Header 2")]
-
-    html_splitter = HTMLSectionSplitter(headers_to_split_on=headers_to_split_on)
-    html_section_splits = html_splitter.split_text(input_text)
-    logger.info(html_section_splits[0])
-    
-    return headers_to_split_on, html_section_splits
